✨ Introduce key listener hook
