+++ src/shared/hooks/ui/useKeyBinds.ts
@@ -1,8 +1,9 @@
 // src/shared/hooks/ui/useKeyBinds.ts
 'use client';
 
-import { useEffect } from 'react';
+import { useCallback, useMemo } from 'react';
 import { KEY_BINDS } from '@/shared/constants';
+import { useKeyListener } from './useKeyListener';
 
 interface KeyBindOptions {
   onPlanner: () => void;
@@ -26,34 +27,25 @@ export function useKeyBinds({
   onCatalog,
   isActive = true,
 }: KeyBindOptions) {
-  useEffect(() => {
-    if (!isActive) return;
-
-    const handlers: Record<string, () => void> = {
+  const handlers = useMemo(
+    () => ({
       [KEY_BINDS.planner]: onPlanner,
       [KEY_BINDS.map]: onMap,
       [KEY_BINDS.budget]: onBudget,
       [KEY_BINDS.newCard]: onNewCard,
       [KEY_BINDS.catalog]: onCatalog,
-    };
+    }),
+    [onPlanner, onMap, onBudget, onNewCard, onCatalog]
+  );
 
-    function handleKeyDown(e: KeyboardEvent) {
-      const active = document.activeElement as HTMLElement | null;
-      const isTextField =
-        active?.tagName === 'INPUT' || active?.tagName === 'TEXTAREA' || active?.isContentEditable;
+  const filter = useCallback((e: KeyboardEvent) => {
+    void e;
+    const active = document.activeElement as HTMLElement | null;
+    const isTextField =
+      active?.tagName === 'INPUT' || active?.tagName === 'TEXTAREA' || active?.isContentEditable;
 
-      if (isTextField) return;
-      const key = e.key.toLowerCase();
-      const handler = handlers[key];
-      if (handler) {
-        e.preventDefault();
-        handler();
-      }
-    }
+    return !isTextField;
+  }, []);
 
-    document.addEventListener('keydown', handleKeyDown);
-    return () => {
-      document.removeEventListener('keydown', handleKeyDown);
-    };
-  }, [isActive, onPlanner, onMap, onBudget, onNewCard, onCatalog]);
+  useKeyListener({ keys: handlers, isActive, filter });
 }
